@@ -58,19 +58,20 @@
     "azure-arm-search": "^1.3.0-preview",
     "azure-arm-sql": "5.6.0",
     "azure-arm-website": "5.7.0",
+    "bluebird": "^3.5.3",
     "body-parser": "1.18.3",
-    "botbuilder": "^4.1.5",
-    "botbuilder-ai": "^4.1.5",
-    "botbuilder-azure": "^4.1.5",
+    "botbuilder": "^4.1.7",
+    "botbuilder-ai": "^4.1.7",
+    "botbuilder-azure": "^4.1.7",
     "botbuilder-choices": "^4.0.0-preview1.2",
-    "botbuilder-dialogs": "^4.1.5",
+    "botbuilder-dialogs": "^4.1.7",
     "botbuilder-prompts": "^4.0.0-preview1.2",
     "botlib": "^0.1.8",
     "chai": "4.2.0",
     "child_process": "^1.0.2",
     "chokidar": "2.0.4",
     "cli-spinner": "^0.2.8",
-    "csv-parse": "4.0.0",
+    "csv-parse": "4.0.1",
     "dotenv-extended": "2.3.0",
     "express": "4.16.4",
     "express-promise-router": "3.0.3",
@@ -104,9 +105,10 @@
     "ts-node": "7.0.1",
     "typedoc": "0.13.0",
     "typedoc-plugin-external-module-name": "^1.1.3",
-    "typedoc-plugin-markdown": "^1.1.18",
-    "typescript": "3.1.6",
+    "typedoc-plugin-markdown": "^1.1.19",
+    "typescript": "3.2.1",
     "url-join": "4.0.0",
+    "vbscript-to-typescript": "^1.0.8",
     "wait-until": "0.0.2",
     "walk-promise": "0.2.0",
     "winston": "3.1.0"
@@ -115,12 +117,12 @@
     "@semantic-release/changelog": "^3.0.1",
     "@semantic-release/commit-analyzer": "^6.1.0",
     "@semantic-release/git": "^7.0.5",
-    "@semantic-release/github": "^5.2.1",
+    "@semantic-release/github": "^5.2.5",
     "@semantic-release/npm": "^5.1.1",
     "@semantic-release/release-notes-generator": "^7.1.4",
     "@types/chai": "4.1.7",
     "@types/mocha": "5.2.5",
-    "@types/sequelize": "4.27.31",
+    "@types/sequelize": "4.27.32",
     "@types/url-join": "0.8.2",
     "@types/winston": "2.4.4",
     "ban-sensitive-files": "1.9.2",
@@ -134,10 +136,10 @@
     "nsp": "3.2.1",
     "pre-git": "3.17.1",
     "prettier-standard": "8.0.1",
-    "semantic-release": "^15.12.0",
+    "semantic-release": "^15.12.4",
     "standard": "12.0.1",
     "travis-deploy-once": "5.0.9",
-    "ts-loader": "^5.3.0",
+    "ts-loader": "^5.3.1",
     "tslint": "^5.11.0",
     "tslint-microsoft-contrib": "^6.0.0"
   },
