@@ -42,8 +42,9 @@ const UrlJoin = require('url-join');
 const Fs = require('fs');
 const WaitUntil = require('wait-until');
 const express = require('express');
+const child_process = require('child_process');
 
-import { IGBCoreService, IGBInstance } from 'botlib';
+import { GBMinInstance, IGBCoreService, IGBInstance } from 'botlib';
 import { GBError } from 'botlib';
 import { IGBPackage } from 'botlib';
 import { AzureSearch } from 'pragmatismo-io-framework';
@@ -51,7 +52,8 @@ import { AzureDeployerService } from '../../azuredeployer.gbapp/services/AzureDe
 import { GuaribasInstance, GuaribasPackage } from '../models/GBModel';
 import { KBService } from './../../kb.gbapp/services/KBService';
 import { GBConfigService } from './GBConfigService';
-import { GBImporter } from './GBImporter';
+import { GBImporter } from './GBImporterService';
+import { GBVMService } from './GBVMService';
 
 /** Deployer service for bots, themes, ai and more. */
 export class GBDeployer {
@@ -66,13 +68,9 @@ export class GBDeployer {
   }
 
   public static getConnectionStringFromInstance(instance: GuaribasInstance) {
-    return `Server=tcp:${
-      instance.storageServer
-    }.database.windows.net,1433;Database=${instance.storageName};User ID=${
+    return `Server=tcp:${instance.storageServer}.database.windows.net,1433;Database=${instance.storageName};User ID=${
       instance.storageUsername
-    };Password=${
-      instance.storagePassword
-    };Trusted_Connection=False;Encrypt=True;Connection Timeout=30;`;
+    };Password=${instance.storagePassword};Trusted_Connection=False;Encrypt=True;Connection Timeout=30;`;
   }
 
   /**
@@ -80,163 +78,152 @@ export class GBDeployer {
    * Performs package deployment in all .gbai or default.
    *
    * */
-  public deployPackages(
-    core: IGBCoreService,
-    server: any,
-    appPackages: IGBPackage[]
-  ) {
+  public deployPackages(core: IGBCoreService, server: any, appPackages: IGBPackage[]) {
     const _this = this;
-    return new Promise((resolve: any, reject: any): any => {
-      let totalPackages = 0;
-      const additionalPath = GBConfigService.get('ADDITIONAL_DEPLOY_PATH');
-      let paths = [GBDeployer.deployFolder];
-      if (additionalPath) {
-        paths = paths.concat(additionalPath.toLowerCase().split(';'));
-      }
-      const botPackages = new Array<string>();
-      const gbappPackages = new Array<string>();
-      let generalPackages = new Array<string>();
-
-      function doIt(path) {
-        const isDirectory = source => Fs.lstatSync(source).isDirectory();
-        const getDirectories = source =>
-          Fs.readdirSync(source)
-            .map(name => Path.join(source, name))
-            .filter(isDirectory);
-
-        const dirs = getDirectories(path);
-        dirs.forEach(element => {
-          if (element.startsWith('.')) {
-            logger.info(`Ignoring ${element}...`);
-          } else {
-            if (element.endsWith('.gbot')) {
-              botPackages.push(element);
-            } else if (element.endsWith('.gbapp')) {
-              gbappPackages.push(element);
-            } else {
-              generalPackages.push(element);
-            }
-          }
-        });
-      }
-
-      logger.info(
-        `Starting looking for packages (.gbot, .gbtheme, .gbkb, .gbapp)...`
-      );
-      paths.forEach(e => {
-        logger.info(`Looking in: ${e}...`);
-        doIt(e);
-      });
-
-      /** Deploys all .gbapp files first. */
-
-      let appPackagesProcessed = 0;
-
-      gbappPackages.forEach(e => {
-        // Skips .gbapp inside deploy folder.
-        if (!e.startsWith('packages')) {
-          logger.info(`Deploying app: ${e}...`);
-          import(e)
-            .then(m => {
-              const p = new m.Package();
-              p.loadPackage(core, core.sequelize);
-              appPackages.push(p);
-              logger.info(`App (.gbapp) deployed: ${e}.`);
-              appPackagesProcessed++;
-            })
-            .catch(err => {
-              logger.error(`Error deploying App (.gbapp): ${e}: ${err}`);
-              appPackagesProcessed++;
-            });
-        } else {
-          appPackagesProcessed++;
+    return new Promise(
+      (resolve: any, reject: any): any => {
+        let totalPackages = 0;
+        const additionalPath = GBConfigService.get('ADDITIONAL_DEPLOY_PATH');
+        let paths = [GBDeployer.deployFolder];
+        if (additionalPath) {
+          paths = paths.concat(additionalPath.toLowerCase().split(';'));
         }
-      });
+        const botPackages = new Array<string>();
+        const gbappPackages = new Array<string>();
+        let generalPackages = new Array<string>();
 
-      WaitUntil()
-        .interval(1000)
-        .times(10)
-        .condition(function(cb) {
-          logger.info(`Waiting for app package deployment...`);
-          cb(appPackagesProcessed == gbappPackages.length);
-        })
-        .done(async result => {
-          logger.info(`App Package deployment done.`);
+        function doIt(path) {
+          const isDirectory = source => Fs.lstatSync(source).isDirectory();
+          const getDirectories = source =>
+            Fs.readdirSync(source)
+              .map(name => Path.join(source, name))
+              .filter(isDirectory);
 
-          try {
-            await core.syncDatabaseStructure();
-          } catch (e) {
-            throw e;
-          }
-
-          /** Deploys all .gbot files first. */
-
-          botPackages.forEach(e => {
-            logger.info(`Deploying bot: ${e}...`);
-            _this.deployBot(e);
-            logger.info(`Bot: ${e} deployed...`);
-          });
-
-          /** Then all remaining generalPackages are loaded. */
-
-          generalPackages = generalPackages.filter(p => !p.endsWith('.git'));
-
-          generalPackages.forEach(filename => {
-            const filenameOnly = Path.basename(filename);
-            logger.info(`Deploying package: ${filename}...`);
-
-            /** Handles apps for general bots - .gbapp must stay out of deploy folder. */
-
-            if (
-              Path.extname(filename) === '.gbapp' ||
-              Path.extname(filename) === '.gblib'
-            ) {
-              /** Themes for bots. */
-            } else if (Path.extname(filename) === '.gbtheme') {
-              server.use('/themes/' + filenameOnly, express.static(filename));
-              logger.info(
-                `Theme (.gbtheme) assets accessible at: ${'/themes/' +
-                  filenameOnly}.`
-              );
-
-              /** Knowledge base for bots. */
-            } else if (Path.extname(filename) === '.gbkb') {
-              server.use(
-                '/kb/' + filenameOnly + '/subjects',
-                express.static(UrlJoin(filename, 'subjects'))
-              );
-              logger.info(
-                `KB (.gbkb) assets accessible at: ${'/kb/' + filenameOnly}.`
-              );
-            } else if (Path.extname(filename) === '.gbui') {
-              // Already Handled
+          const dirs = getDirectories(path);
+          dirs.forEach(element => {
+            if (element.startsWith('.')) {
+              logger.info(`Ignoring ${element}...`);
             } else {
-              /** Unknown package format. */
-              const err = new Error(`Package type not handled: ${filename}.`);
-              reject(err);
-            }
-            totalPackages++;
-          });
-
-          WaitUntil()
-            .interval(100)
-            .times(5)
-            .condition(function(cb) {
-              logger.info(`Waiting for package deployment...`);
-              cb(totalPackages == generalPackages.length);
-            })
-            .done(function(result) {
-              if (botPackages.length === 0) {
-                logger.info(
-                  'No external packages to load, please use ADDITIONAL_DEPLOY_PATH to point to a .gbai package folder.'
-                );
+              if (element.endsWith('.gbot')) {
+                botPackages.push(element);
+              } else if (element.endsWith('.gbapp')) {
+                gbappPackages.push(element);
               } else {
-                logger.info(`Package deployment done.`);
+                generalPackages.push(element);
               }
-              resolve();
-            });
+            }
+          });
+        }
+
+        logger.info(`Starting looking for packages (.gbot, .gbtheme, .gbkb, .gbapp)...`);
+        paths.forEach(e => {
+          logger.info(`Looking in: ${e}...`);
+          doIt(e);
         });
-    });
+
+        /** Deploys all .gbapp files first. */
+
+        let appPackagesProcessed = 0;
+
+        gbappPackages.forEach(e => {
+          // Skips .gbapp inside deploy folder.
+          if (!e.startsWith('packages')) {
+            logger.info(`Deploying app: ${e}...`);
+            import(e)
+              .then(m => {
+                const p = new m.Package();
+                p.loadPackage(core, core.sequelize);
+                appPackages.push(p);
+                logger.info(`App (.gbapp) deployed: ${e}.`);
+                appPackagesProcessed++;
+              })
+              .catch(err => {
+                logger.error(`Error deploying App (.gbapp): ${e}: ${err}`);
+                appPackagesProcessed++;
+              });
+          } else {
+            appPackagesProcessed++;
+          }
+        });
+
+        WaitUntil()
+          .interval(1000)
+          .times(10)
+          .condition(function(cb) {
+            logger.info(`Waiting for app package deployment...`);
+            cb(appPackagesProcessed == gbappPackages.length);
+          })
+          .done(async result => {
+            logger.info(`App Package deployment done.`);
+
+            try {
+              await core.syncDatabaseStructure();
+            } catch (e) {
+              throw e;
+            }
+
+            /** Deploys all .gbot files first. */
+
+            botPackages.forEach(e => {
+              if (e !== 'packages\\boot.gbot') {
+                logger.info(`Deploying bot: ${e}...`);
+                _this.deployBot(e);
+                logger.info(`Bot: ${e} deployed...`);
+              }
+            });
+
+            /** Then all remaining generalPackages are loaded. */
+
+            generalPackages = generalPackages.filter(p => !p.endsWith('.git'));
+
+            generalPackages.forEach(filename => {
+              const filenameOnly = Path.basename(filename);
+              logger.info(`Deploying package: ${filename}...`);
+
+              /** Handles apps for general bots - .gbapp must stay out of deploy folder. */
+
+              if (Path.extname(filename) === '.gbapp' || Path.extname(filename) === '.gblib') {
+                /** Themes for bots. */
+              } else if (Path.extname(filename) === '.gbtheme') {
+                server.use('/themes/' + filenameOnly, express.static(filename));
+                logger.info(`Theme (.gbtheme) assets accessible at: ${'/themes/' + filenameOnly}.`);
+
+                /** Knowledge base for bots. */
+              } else if (Path.extname(filename) === '.gbkb') {
+                server.use('/kb/' + filenameOnly + '/subjects', express.static(UrlJoin(filename, 'subjects')));
+                logger.info(`KB (.gbkb) assets accessible at: ${'/kb/' + filenameOnly}.`);
+              } else if (Path.extname(filename) === '.gbui') {
+                // Already Handled
+              } else if (Path.extname(filename) === '.gbdialog') {
+                // Already Handled
+              } else {
+                /** Unknown package format. */
+                const err = new Error(`Package type not handled: ${filename}.`);
+                reject(err);
+              }
+              totalPackages++;
+            });
+
+            WaitUntil()
+              .interval(100)
+              .times(5)
+              .condition(function(cb) {
+                logger.info(`Waiting for package deployment...`);
+                cb(totalPackages == generalPackages.length);
+              })
+              .done(function(result) {
+                if (botPackages.length === 0) {
+                  logger.info(
+                    'No external packages to load, please use ADDITIONAL_DEPLOY_PATH to point to a .gbai package folder.'
+                  );
+                } else {
+                  logger.info(`Package deployment done.`);
+                }
+                resolve();
+              });
+          });
+      }
+    );
   }
 
   /**
@@ -246,22 +233,19 @@ export class GBDeployer {
   public async deployBot(localPath: string): Promise<IGBInstance> {
     const packageType = Path.extname(localPath);
     const packageName = Path.basename(localPath);
-    const instance = await this.importer.importIfNotExistsBotPackage(
-      packageName,
-      localPath
-    );
+    const instance = await this.importer.importIfNotExistsBotPackage(null, packageName, localPath);
+
     return instance;
   }
 
-  public async deployPackageToStorage(
-    instanceId: number,
-    packageName: string
-  ): Promise<GuaribasPackage> {
+  public async deployPackageToStorage(instanceId: number, packageName: string): Promise<GuaribasPackage> {
     return GuaribasPackage.create({
       packageName: packageName,
       instanceId: instanceId
     });
   }
+
+  public deployScriptToStorage(instanceId: number, localPath: string) {}
 
   public deployTheme(localPath: string) {
     // DISABLED: Until completed, "/ui/public".
@@ -275,7 +259,7 @@ export class GBDeployer {
     //   })
   }
 
-  public async deployPackageFromLocalPath(localPath: string) {
+  public async deployPackageFromLocalPath(min: GBMinInstance, localPath: string) {
     const packageType = Path.extname(localPath);
 
     switch (packageType) {
@@ -293,10 +277,12 @@ export class GBDeployer {
       case '.gbui':
         break;
 
+      case '.gbdialog':
+        const vm = new GBVMService();
+        return vm.loadJS(localPath, min, this.core, this, localPath);
+
       default:
-        const err = GBError.create(
-          `GuaribasBusinessError: Unknow package type: ${packageType}.`
-        );
+        const err = GBError.create(`GuaribasBusinessError: Unknown package type: ${packageType}.`);
         Promise.reject(err);
         break;
     }
@@ -324,10 +310,11 @@ export class GBDeployer {
       case '.gbui':
         break;
 
+      case '.gbdialog':
+        break;
+
       default:
-        const err = GBError.create(
-          `GuaribasBusinessError: Unknown package type: ${packageType}.`
-        );
+        const err = GBError.create(`GuaribasBusinessError: Unknown package type: ${packageType}.`);
         Promise.reject(err);
         break;
     }
@@ -353,13 +340,7 @@ export class GBDeployer {
       }
     }
 
-    await search.createDataSource(
-      dsName,
-      dsName,
-      'GuaribasQuestion',
-      'azuresql',
-      connectionString
-    );
+    await search.createDataSource(dsName, dsName, 'GuaribasQuestion', 'azuresql', connectionString);
 
     try {
       await search.deleteIndex();
@@ -369,37 +350,22 @@ export class GBDeployer {
         throw err;
       }
     }
-    await search.createIndex(
-      AzureDeployerService.getKBSearchSchema(instance.searchIndex),
-      dsName
-    );
+    await search.createIndex(AzureDeployerService.getKBSearchSchema(instance.searchIndex), dsName);
   }
 
-  public async getPackageByName(
-    instanceId: number,
-    packageName: string
-  ): Promise<GuaribasPackage> {
+  public async getPackageByName(instanceId: number, packageName: string): Promise<GuaribasPackage> {
     const where = { packageName: packageName, instanceId: instanceId };
     return GuaribasPackage.findOne({
       where: where
     });
   }
 
-  /**
-   *
-   * Hot deploy processing.
-   *
-   */
-  public async scanBootPackage() {
-    const deployFolder = 'packages';
-    const bootPackage = GBConfigService.get('BOOT_PACKAGE');
-
-    if (bootPackage === 'none') {
-      return Promise.resolve(true);
-    } else {
-      return this.deployPackageFromLocalPath(
-        UrlJoin(deployFolder, bootPackage)
-      );
+  public installDefaultGBUI() {
+    const root = 'packages/default.gbui';
+    if (!Fs.existsSync(`${root}/build`)) {
+      Fs.writeFileSync(`${root}/.env`, 'SKIP_PREFLIGHT_CHECK=true');
+      child_process.execSync('npm install', { cwd: root });
+      child_process.execSync('npm run build', { cwd: root });
     }
   }
 }
