@@ -43,21 +43,9 @@ const logger = require('../../../src/logger');
 const request = require('request-promise-native');
 const AuthenticationContext = require('adal-node').AuthenticationContext;
 
-import {
-  AutoSaveStateMiddleware,
-  BotFrameworkAdapter,
-  ConversationState,
-  MemoryStorage,
-  UserState
-} from 'botbuilder';
+import { AutoSaveStateMiddleware, BotFrameworkAdapter, ConversationState, MemoryStorage, UserState } from 'botbuilder';
 
-import {
-  GBMinInstance,
-  IGBAdminService,
-  IGBConversationalService,
-  IGBCoreService,
-  IGBPackage
-} from 'botlib';
+import { GBMinInstance, IGBAdminService, IGBConversationalService, IGBCoreService, IGBPackage } from 'botlib';
 import { GBAnalyticsPackage } from '../../analytics.gblib';
 import { GBCorePackage } from '../../core.gbapp';
 import { GBCustomerSatisfactionPackage } from '../../customer-satisfaction.gbapp';
@@ -88,7 +76,7 @@ export class GBMinService {
     core: IGBCoreService,
     conversationalService: IGBConversationalService,
     adminService: IGBAdminService,
-    deployer: GBDeployer,
+    deployer: GBDeployer
   ) {
     this.core = core;
     this.conversationalService = conversationalService;
@@ -108,17 +96,16 @@ export class GBMinService {
    * */
 
   public async buildMin(
+    bootInstance: GuaribasInstance,
     server: any,
     appPackages: IGBPackage[],
     instances: GuaribasInstance[],
+    deployer: GBDeployer
   ): Promise<GBMinInstance> {
     // Serves default UI on root address '/'.
 
     const uiPackage = 'default.gbui';
-    server.use(
-      '/',
-      express.static(UrlJoin(GBDeployer.deployFolder, uiPackage, 'build')),
-    );
+    server.use('/', express.static(UrlJoin(GBDeployer.deployFolder, uiPackage, 'build')));
 
     Promise.all(
       instances.map(async instance => {
@@ -133,7 +120,11 @@ export class GBMinService {
           (async () => {
             // Returns the instance object to clients requesting bot info.
 
-            const botId = req.params.botId;
+            let botId = req.params.botId;
+            if (botId === '[default]') {
+              botId = bootInstance.botId;
+            }
+
             const instance = await this.core.loadInstance(botId);
 
             if (instance) {
@@ -152,8 +143,8 @@ export class GBMinService {
                   speechToken: speechToken,
                   conversationId: webchatToken.conversationId,
                   authenticatorTenant: instance.authenticatorTenant,
-                  authenticatorClientId: instance.authenticatorClientId,
-                }),
+                  authenticatorClientId: instance.authenticatorClientId
+                })
               );
             } else {
               const error = `Instance not found: ${botId}.`;
@@ -165,9 +156,11 @@ export class GBMinService {
 
         // Build bot adapter.
 
-        const { min, adapter, conversationState } = await this.buildBotAdapter(
-          instance,
-        );
+        const { min, adapter, conversationState } = await this.buildBotAdapter(instance);
+
+        // Install default VBA module.
+
+        deployer.deployPackageFromLocalPath(min, 'packages/default.gbdialog');
 
         // Call the loadBot context.activity for all packages.
 
@@ -177,32 +170,17 @@ export class GBMinService {
 
         const url = `/api/messages/${instance.botId}`;
         server.post(url, async (req, res) => {
-          return this.receiver(
-            adapter,
-            req,
-            res,
-            conversationState,
-            min,
-            instance,
-            appPackages,
-          );
+          return await this.receiver(adapter, req, res, conversationState, min, instance, appPackages);
         });
-        logger.info(
-          `GeneralBots(${instance.engineName}) listening on: ${url}.`,
-        );
+        logger.info(`GeneralBots(${instance.engineName}) listening on: ${url}.`);
 
         // Serves individual URL for each bot user interface.
 
         const uiUrl = `/${instance.botId}`;
-        server.use(
-          uiUrl,
-          express.static(UrlJoin(GBDeployer.deployFolder, uiPackage, 'build')),
-        );
+        server.use(uiUrl, express.static(UrlJoin(GBDeployer.deployFolder, uiPackage, 'build')));
 
         logger.info(`Bot UI ${uiPackage} accessible at: ${uiUrl}.`);
-        const state = `${instance.instanceId}${Math.floor(
-          Math.random() * 1000000000,
-        )}`;
+        const state = `${instance.instanceId}${Math.floor(Math.random() * 1000000000)}`;
 
         // Clients get redirected here in order to create an OAuth authorize url and redirect them to AAD.
         // There they will authenticate and give their consent to allow this app access to
@@ -211,13 +189,11 @@ export class GBMinService {
           let authorizationUrl = UrlJoin(
             min.instance.authenticatorAuthorityHostUrl,
             min.instance.authenticatorTenant,
-            '/oauth2/authorize',
+            '/oauth2/authorize'
           );
           authorizationUrl = `${authorizationUrl}?response_type=code&client_id=${
             min.instance.authenticatorClientId
-          }&redirect_uri=${min.instance.botEndpoint}/${
-            min.instance.botId
-          }/token`;
+          }&redirect_uri=${min.instance.botEndpoint}/${min.instance.botId}/token`;
 
           res.redirect(authorizationUrl);
         });
@@ -227,23 +203,16 @@ export class GBMinService {
         // access token that can be used to access the user owned resource.
 
         server.get(`/${min.instance.botId}/token`, async (req, res) => {
-          const state = await min.adminService.getValue(
-            min.instance.instanceId,
-            'AntiCSRFAttackState',
-          );
+          const state = await min.adminService.getValue(min.instance.instanceId, 'AntiCSRFAttackState');
 
           if (req.query.state !== state) {
-            const msg =
-              'WARNING: state field was not provided as anti-CSRF token';
+            const msg = 'WARNING: state field was not provided as anti-CSRF token';
             logger.error(msg);
             throw new Error(msg);
           }
 
           const authenticationContext = new AuthenticationContext(
-            UrlJoin(
-              min.instance.authenticatorAuthorityHostUrl,
-              min.instance.authenticatorTenant,
-            ),
+            UrlJoin(min.instance.authenticatorAuthorityHostUrl, min.instance.authenticatorTenant)
           );
 
           const resource = 'https://graph.microsoft.com';
@@ -260,48 +229,17 @@ export class GBMinService {
                 logger.error(msg);
                 res.send(msg);
               } else {
-                await this.adminService.setValue(
-                  instance.instanceId,
-                  'refreshToken',
-                  token.refreshToken,
-                );
-                await this.adminService.setValue(
-                  instance.instanceId,
-                  'accessToken',
-                  token.accessToken,
-                );
-                await this.adminService.setValue(
-                  instance.instanceId,
-                  'expiresOn',
-                  token.expiresOn.toString(),
-                );
-                await this.adminService.setValue(
-                  instance.instanceId,
-                  'AntiCSRFAttackState',
-                  null,
-                );
+                await this.adminService.setValue(instance.instanceId, 'refreshToken', token.refreshToken);
+                await this.adminService.setValue(instance.instanceId, 'accessToken', token.accessToken);
+                await this.adminService.setValue(instance.instanceId, 'expiresOn', token.expiresOn.toString());
+                await this.adminService.setValue(instance.instanceId, 'AntiCSRFAttackState', null);
 
                 res.redirect(min.instance.botEndpoint);
               }
-            },
+            }
           );
         });
-
-        // Setups handlers.
-        // send: function (context.activity, next) {
-        //   logger.info(
-        //     `[SND]: ChannelID: ${context.activity.address.channelId}, ConversationID: ${context.activity.address.conversation},
-        //      Type: ${context.activity.type}              `)
-        //   this.core.createMessage(
-        //     this.min.conversation,
-        //     this.min.conversation.startedBy,
-        //     context.activity.source,
-        //     (data, err) => {
-        //       logger.info(context.activity.source)
-        //     }
-        //   )
-        //   next()
-      }),
+      })
     );
   }
 
@@ -316,8 +254,8 @@ export class GBMinService {
       url: 'https://directline.botframework.com/v3/directline/tokens/generate',
       method: 'POST',
       headers: {
-        Authorization: `Bearer ${instance.webchatKey}`,
-      },
+        Authorization: `Bearer ${instance.webchatKey}`
+      }
     };
 
     try {
@@ -344,8 +282,8 @@ export class GBMinService {
       url: 'https://westus.api.cognitive.microsoft.com/sts/v1.0/issueToken',
       method: 'POST',
       headers: {
-        'Ocp-Apim-Subscription-Key': instance.speechKey,
-      },
+        'Ocp-Apim-Subscription-Key': instance.speechKey
+      }
     };
 
     try {
@@ -359,7 +297,7 @@ export class GBMinService {
   private async buildBotAdapter(instance: any) {
     const adapter = new BotFrameworkAdapter({
       appId: instance.marketplaceId,
-      appPassword: instance.marketplacePassword,
+      appPassword: instance.marketplacePassword
     });
 
     const storage = new MemoryStorage();
@@ -377,8 +315,10 @@ export class GBMinService {
     min.conversationalService = this.conversationalService;
     min.adminService = this.adminService;
     min.instance = await this.core.loadInstance(min.botId);
+    min.cbMap = {};
     min.userProfile = conversationState.createProperty('userProfile');
     const dialogState = conversationState.createProperty('dialogState');
+
     min.dialogs = new DialogSet(dialogState);
     min.dialogs.add(new TextPrompt('textPrompt'));
 
@@ -395,7 +335,7 @@ export class GBMinService {
       GBKBPackage,
       GBAnalyticsPackage,
       GBCustomerSatisfactionPackage,
-      GBWhatsappPackage,
+      GBWhatsappPackage
     ].forEach(sysPackage => {
       const p = Object.create(sysPackage.prototype) as IGBPackage;
       p.loadBot(min);
@@ -417,17 +357,18 @@ export class GBMinService {
   /**
    * Bot Service hook method.
    */
-  private receiver(
+  private async receiver(
     adapter: BotFrameworkAdapter,
     req: any,
     res: any,
     conversationState: ConversationState,
     min: any,
     instance: any,
-    appPackages: any[],
+    appPackages: any[]
   ) {
-    return adapter.processActivity(req, res, async context => {
-      const state = conversationState.get(context);
+    return await adapter.processActivity(req, res, async context => {
+      // Get loaded user state
+      const state = await conversationState.get(context);
       const step = await min.dialogs.createContext(context, state);
       step.context.activity.locale = 'en-US'; // TODO: Make dynamic.
 
@@ -439,22 +380,20 @@ export class GBMinService {
             instanceId: instance.instanceId,
             botId: instance.botId,
             theme: instance.theme ? instance.theme : 'default.gbtheme',
-            secret: instance.webchatKey,
+            secret: instance.webchatKey
           });
           user.loaded = true;
           user.subjects = [];
+          user.cb = null;
           await min.userProfile.set(step.context, user);
         }
 
         logger.info(
-          `User>: ${context.activity.text} (${context.activity.type}, ${
-            context.activity.name
-          }, ${context.activity.channelId}, {context.activity.value})`,
+          `User>: ${context.activity.text} (${context.activity.type}, ${context.activity.name}, ${
+            context.activity.channelId
+          }, {context.activity.value})`
         );
-        if (
-          context.activity.type === 'conversationUpdate' &&
-          context.activity.membersAdded.length > 0
-        ) {
+        if (context.activity.type === 'conversationUpdate' && context.activity.membersAdded.length > 0) {
           const member = context.activity.membersAdded[0];
           if (member.name === 'GeneralBots') {
             logger.info(`Bot added to conversation, starting chat...`);
@@ -471,23 +410,29 @@ export class GBMinService {
           // Processes messages.
         } else if (context.activity.type === 'message') {
           // Checks for /admin request.
-
-          if (context.activity.text === 'admin') {
+          if (context.activity.text === 'alpha-vba') {
+            min.sandbox.context = context;
+            min.sandbox.step = step;
+            min.sandbox['bot'].bind(min.sandbox);
+            await min.sandbox['bot']();
+          } else if (context.activity.text === 'admin') {
             await step.beginDialog('/admin');
 
             // Checks for /menu JSON signature.
           } else if (context.activity.text.startsWith('{"title"')) {
             await step.beginDialog('/menu', {
-              data: JSON.parse(context.activity.text),
+              data: JSON.parse(context.activity.text)
             });
 
             // Otherwise, continue to the active dialog in the stack.
           } else {
+            const user = await min.userProfile.get(context, {});
+
             if (step.activeDialog) {
               await step.continueDialog();
             } else {
               await step.beginDialog('/answer', {
-                query: context.activity.text,
+                query: context.activity.text
               });
             }
           }
@@ -504,18 +449,18 @@ export class GBMinService {
             await step.beginDialog('/menu');
           } else if (context.activity.name === 'giveFeedback') {
             await step.beginDialog('/feedback', {
-              fromMenu: true,
+              fromMenu: true
             });
           } else if (context.activity.name === 'showFAQ') {
             await step.beginDialog('/faq');
           } else if (context.activity.name === 'answerEvent') {
             await step.beginDialog('/answerEvent', {
               questionId: (context.activity as any).data,
-              fromFaq: true,
+              fromFaq: true
             });
           } else if (context.activity.name === 'quality') {
             await step.beginDialog('/quality', {
-              score: (context.activity as any).data,
+              score: (context.activity as any).data
             });
           } else if (context.activity.name === 'updateToken') {
             const token = (context.activity as any).data;
@@ -524,13 +469,12 @@ export class GBMinService {
             await step.continueDialog();
           }
         }
+        await conversationState.saveChanges(context, true);
       } catch (error) {
         const msg = `ERROR: ${error.message} ${error.stack ? error.stack : ''}`;
         logger.error(msg);
 
-        await step.context.sendActivity(
-          Messages[step.context.activity.locale].very_sorry_about_error,
-        );
+        await step.context.sendActivity(Messages[step.context.activity.locale].very_sorry_about_error);
         await step.beginDialog('/ask', { isReturning: true });
       }
     });
