@@ -40,25 +40,17 @@
 const logger = require('./logger');
 const express = require('express');
 const bodyParser = require('body-parser');
-const opn = require('opn');
-
 import { IGBInstance, IGBPackage } from 'botlib';
-import { GBAdminPackage } from '../packages/admin.gbapp/index';
 import { GBAdminService } from '../packages/admin.gbapp/services/GBAdminService';
-import { GBAnalyticsPackage } from '../packages/analytics.gblib';
 import { AzureDeployerService } from '../packages/azuredeployer.gbapp/services/AzureDeployerService';
-import { GBCorePackage } from '../packages/core.gbapp';
 import { GuaribasInstance } from '../packages/core.gbapp/models/GBModel';
 import { GBConfigService } from '../packages/core.gbapp/services/GBConfigService';
 import { GBConversationalService } from '../packages/core.gbapp/services/GBConversationalService';
 import { GBCoreService } from '../packages/core.gbapp/services/GBCoreService';
 import { GBDeployer } from '../packages/core.gbapp/services/GBDeployer';
-import { GBImporter } from '../packages/core.gbapp/services/GBImporter';
+import { GBImporter } from '../packages/core.gbapp/services/GBImporterService';
 import { GBMinService } from '../packages/core.gbapp/services/GBMinService';
-import { GBCustomerSatisfactionPackage } from '../packages/customer-satisfaction.gbapp';
-import { GBKBPackage } from '../packages/kb.gbapp';
-import { GBSecurityPackage } from '../packages/security.gblib';
-import { GBWhatsappPackage } from './../packages/whatsapp.gblib/index';
+import { GBVMService } from '../packages/core.gbapp/services/GBVMService';
 
 const appPackages = new Array<IGBPackage>();
 
@@ -66,13 +58,11 @@ const appPackages = new Array<IGBPackage>();
  * General Bots open-core entry point.
  */
 export class GBServer {
-
   /**
    *  Program entry-point.
    */
 
   public static run() {
-
     logger.info(`The Bot Server is in STARTING mode...`);
 
     // Creates a basic HTTP server that will serve several URL, one for each
@@ -104,135 +94,51 @@ export class GBServer {
           // Ensures cloud / on-premises infrastructure is setup.
 
           logger.info(`Establishing a development local proxy (ngrok)...`);
-          const proxyAddress = await core.ensureProxy(port);
-
-          const deployer = new GBDeployer(core, new GBImporter(core));
-          const azureDeployer = new AzureDeployerService(deployer);
-
-          try {
-            await core.initDatabase();
-          } catch (error) {
-            logger.info(`Deploying cognitive infrastructure (on the cloud / on premises)...`);
-            try {
-              bootInstance = await azureDeployer.deployFarm(proxyAddress);
-            } catch (error) {
-              logger.warn(
-                'In case of error, please cleanup any infrastructure objects created during this procedure and .env before running again.'
-              );
-              throw error;
-            }
-            core.writeEnv(bootInstance);
-            logger.info(`File .env written, starting General Bots...`);
-            GBConfigService.init();
-
-            await core.initDatabase();
-          }
-
-          // TODO: Get .gb* templates from GitHub and download do additional deploy folder.
-
-          // Check admin password.
-
-          const conversationalService = new GBConversationalService(core);
-          const adminService = new GBAdminService(core);
-          const password = GBConfigService.get('ADMIN_PASS');
-
-          if (!GBAdminService.StrongRegex.test(password)) {
-            throw new Error(
-              'Please, define a really strong password in ADMIN_PASS environment variable before running the server.'
-            );
-          }
-
-          // NOTE: the semicolon is necessary before this line.
-          // Loads all system packages.
-
-          [
-            GBAdminPackage,
-            GBAnalyticsPackage,
-            GBCorePackage,
-            GBSecurityPackage,
-            GBKBPackage,
-            GBCustomerSatisfactionPackage,
-            GBWhatsappPackage
-          ].forEach(e => {
-            logger.info(`Loading sys package: ${e.name}...`);
-            const p = Object.create(e.prototype) as IGBPackage;
-            p.loadPackage(core, core.sequelize);
-          });
-
-          // Loads all bot instances from object storage, if it's formatted.
-
-          logger.info(`Loading instances from storage...`);
-          let instances: GuaribasInstance[];
-          try {
-            instances = await core.loadInstances();
-            const instance = instances[0];
-
-            if (process.env.NODE_ENV === 'development') {
-              logger.info(`Updating bot endpoint to local reverse proxy (ngrok)...`);
-
-              await azureDeployer.updateBotProxy(
-                instance.botId,
-                instance.botId,
-                `${proxyAddress}/api/messages/${instance.botId}`
-              );
-            }
-          } catch (error) {
-            if (error.parent.code === 'ELOGIN') {
-              const group = GBConfigService.get('CLOUD_GROUP');
-              const serverName = GBConfigService.get('STORAGE_SERVER').split(
-                '.database.windows.net'
-              )[0];
-              await azureDeployer.openStorageFirewall(group, serverName);
-            } else {
-              // Check if storage is empty and needs formatting.
-
-              const isInvalidObject =
-                error.parent.number == 208 || error.parent.errno == 1; // MSSQL or SQLITE.
-
-              if (isInvalidObject) {
-                if (GBConfigService.get('STORAGE_SYNC') != 'true') {
-                  throw new Error(`Operating storage is out of sync or there is a storage connection error. Try setting STORAGE_SYNC to true in .env file. Error: ${
-                    error.message
-                  }.`);
-                } else {
-                  logger.info(
-                    `Storage is empty. After collecting storage structure from all .gbapps it will get synced.`
-                  );
-                }
-              } else {
-                throw new Error(`Cannot connect to operating storage: ${error.message}.`);
-              }
-            }
-          }
-
-          // Deploy packages and format object store according to .gbapp storage models.
+         
+          const proxyAddress: string = await core.ensureProxy(port);
 
           logger.info(`Deploying packages...`);
+          const importer: GBImporter = new GBImporter(core);
+          const deployer: GBDeployer = new GBDeployer(core, importer);
+          const azureDeployer: AzureDeployerService = new AzureDeployerService(deployer);
+          const adminService: GBAdminService = new GBAdminService(core);
+          const conversationalService: GBConversationalService = new GBConversationalService(core);
+          core.ensureAdminIsSecured();
+
+          let bootInstance: IGBInstance = null;
+          try {
+            await core.initStorage();
+          } catch (error) {
+            bootInstance = await core.createBootInstance(core, azureDeployer, proxyAddress);
+            await core.initStorage();
+          }
+
+          await core.loadSysPackages(core);
+          await core.checkStorage(azureDeployer);
           await deployer.deployPackages(core, server, appPackages);
 
-          // If instances is undefined here it's because storage has been formatted.
-          // Load all instances from .gbot found on deploy package directory.
-          if (!instances) {
-            const saveInstance = new GuaribasInstance(bootInstance);
-            await saveInstance.save();
-            instances = await core.loadInstances();
-          }
-
-          // Setup server dynamic (per bot instance) resources and listeners.
-
           logger.info(`Publishing instances...`);
-          const minService = new GBMinService(
-            core,
-            conversationalService,
-            adminService,
-            deployer
+          const packageInstance = await importer.importIfNotExistsBotPackage(
+            GBConfigService.get('CLOUD_GROUP'),
+            'boot.gbot',
+            'packages/boot.gbot'
           );
-          await minService.buildMin(server, appPackages, instances);
-          logger.info(`The Bot Server is in RUNNING mode...`);
-
-          if (process.env.NODE_ENV === 'development') {
-            opn('http://localhost:4242');
+          const fullInstance = Object.assign(packageInstance, bootInstance);
+          await core.saveInstance(fullInstance);
+          let instances: GuaribasInstance[] = await core.loadAllInstances(core, azureDeployer, proxyAddress);
+          instances = await core.ensureInstances(instances, bootInstance, core);
+          if(!bootInstance) {
+            bootInstance = instances[0];
           }
+
+          const minService: GBMinService = new GBMinService(core, conversationalService, adminService, deployer);
+          await minService.buildMin(bootInstance, server, appPackages, instances, deployer);
+
+          logger.info(`Preparing default.gbui (it may take some additional time for the first time)...`);
+          deployer.installDefaultGBUI();
+
+          logger.info(`The Bot Server is in RUNNING mode...`);
+          core.openBrowserInDevelopment();
 
           return core;
         } catch (err) {
@@ -245,5 +151,11 @@ export class GBServer {
 }
 
 // First line to run.
+
+// const path = 'packages/default.gbdialog';
+// const file = 'bot.vbs';
+// const source =(path + '/' + file);
+// let s = new GBVMService();
+// s.run(source, path, null, null, null)
 
 GBServer.run();
